refactor(racing): 이동 시도 횟수 정규화 함수를 racingManager로 이동
함수의 목적상 racingManager의 메서드가 되는 것이 더 적절하다고 판단하여 racing 클래스에서 racingManager 클래스로 이동
