+++ src/Racing.js
@@ -1,5 +1,4 @@
 import { Console } from '@woowacourse/mission-utils';
-import Validation from './Validation.js';
 
 class Racing {
   #totalRounds;
@@ -62,16 +61,6 @@ class Racing {
     const winnersNameToString = winnersName.join(', ');
 
     Console.print(`최종 우승자 : ${winnersNameToString}`);
-  }
-
-  static normalizeTotalRounds(totalRounds) {
-    const parsedTotalRounds = this.parseTotalRounds(totalRounds);
-    const validTotalRounds = Validation.totalRounds(parsedTotalRounds);
-    return validTotalRounds;
-  }
-
-  static parseTotalRounds(totalRounds) {
-    return Number(totalRounds);
   }
 }
